Double-Metaphone implementation (shaky)
This was a real monster. I found some C++ code implementing the
algorithm in a mash of sequential case code, and decided to try
and translate it into a bunch of ordered regexes. Not the best
idea I've ever had.

It now works, in the sense that it passes all my available test
cases at the moment. I would not want to trust it, though, especially
for some of the odder names and non-english stuff.