Match Rating Approach Implementation
This was initially meant to be a bit of light relief after the
Double-Metaphone headache. I saw the MRA encoding scheme and added
it quickly. Then I looked at the comparison scheme and realised
that, being phonetic in origin rather than string comparators, that
would be suited to inclusion in this library.

So that's why I've now added 4 functions. One is the 'easy' match
rating coding scheme, one is the similarity metric for strings which
uses it, one is the binary function of *that* metric against a threshold
table and the other is a helper function.