Fuzzy Soundex implementation
The Fuzzy Soundex algorithm was a bit disappointing. I was expecting
something with properly fuzzy coding rules or something, rather than
another variant on 'n-grams before lookup'. The substitutions also
look poor compared to, e.g. Metaphone.

Along with the encoding function, there is a comparison function for
creating a similarity measure between strings. This was vaguely interesting
if obvious.

The paper giving the Fuzzy implementation included mention of code
shifts -- removing trailing or leading characters, but it's not clear
to me how these might be included in the function.