Caverphone; Regex Fixes
Testing the caverphone implementation brought me to realise that
something weird was happening with my regexes across the board.

I eventually found out this was an issue with promotion of strings
in arrays (see https://github.com/JuliaLang/julia/issues/15069)

Once the fix/workaround was found, it was simple enough to apply it.
Driven by the failure of previous test cases, I created an expanded
set for double_metaphone and found a few bugs in that implementation
along the way (and fixed them).