code_cluster implementation
The canopy clustering here uses Phonix rather than Fuzzy Soundex
as the default because it gives generally more sensible-looking
results for the test cases I used.

Also in this commit, I separated the string-encoding front end of
code_similarity from the actual comparison code, which may be useful
elsewhere as well as where I used it in code_cluster.

It seems to be an arse to test vector equality, the default comparator
at least seems to fall over.