Streamlined table_lookup_plural
Also in this commit, several bugs were found with editex() and fixed.

Notably:
 - table_lookup_plural wasn't being called (table_lookup instead, leading
 to errors for 'c'/'p'/'z' distances.
 - the terminal case 1,1 was wrong. This is a bit confusing because the
original alg definition used 0-indexing (like a sane language), but it
seems like the definition given must not work exactly, as either format
sees it indexing outside the bounds of the strings s or t. Presumably
the length input is meant to actually be the maximum index from 0.