Syllable-counting
The new syllable counting function took some time to implement
mostly because it's easy to think of new exceptions to the rules
that seem to break them. Also, you can go crazy trying to decide
if 'area' is two or three syllables long (I settled on two).

Hopefully this function is general enough that it can be used in
other languages, but mostly it's a sane default for English.